16th homework

+++ src/main/java/com/nataliia/servlet/UserServlet.java
@@ -3,6 +3,7 @@ package com.nataliia.servlet;
 import com.nataliia.dao.UserDao;
 import com.nataliia.exceptions.UserNotFoundException;
 import com.nataliia.model.User;
+import org.apache.log4j.Logger;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -10,11 +11,13 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 import java.io.IOException;
 
 @WebServlet(value = "/user")
 public class UserServlet extends HttpServlet {
     private UserDao userDao = new UserDao();
+    private static final Logger logger = Logger.getLogger(AdminServlet.class);
 
     public void setUserDao(UserDao userDao) {
         this.userDao = userDao;
@@ -28,9 +31,27 @@ public class UserServlet extends HttpServlet {
         } catch (NumberFormatException e) {
             throw new UserNotFoundException();
         }
-        User user = userDao.getUser(id).orElseThrow(UserNotFoundException::new);
-        req.setAttribute("user", user);
-        RequestDispatcher dispatcher = getServletContext().getRequestDispatcher("/editUser.jsp");
+
+        HttpSession session = req.getSession();
+        String role = (String) session.getAttribute("role");
+
+        if ("admin".equals(role)) {
+            User user = userDao.getUser(id).orElseThrow(UserNotFoundException::new);
+            logger.debug(user.getName()+ " asked for user information using ID " + "as" + user.getRole());
+            req.setAttribute("user", user);
+        }
+
+        String urlToRedirect;
+        if ("admin".equals(role)) {
+            logger.debug("Admin with ID="+ session.getAttribute("userId")+ " edits user");
+            urlToRedirect = "/editUser.jsp";
+        } else {
+            req.setAttribute("message", "Ошибка. Войдите в систему снова.");
+            logger.debug("Access error");
+            urlToRedirect = "/index.jsp";
+        }
+
+        RequestDispatcher dispatcher = getServletContext().getRequestDispatcher(urlToRedirect);
         dispatcher.forward(req, resp);
     }
 
@@ -46,25 +67,32 @@ public class UserServlet extends HttpServlet {
         } else if ("put".equals(action)) {
             doPut(req, resp);
         } else {
-            String name = req.getParameter("name");
-            String password = req.getParameter("password");
-            userDao.addUser(new User(name, password));
-
-            resp.sendRedirect("/");
+            resp.sendRedirect("/index.jsp");
         }
     }
 
     @Override
     protected void doDelete(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
-        Long id;
-        try {
-            id = Long.parseLong(req.getParameter("id"));
-        } catch (NumberFormatException e) {
-            throw new UserNotFoundException();
-        }
-        boolean isDeleted = userDao.deleteUser(id);
+        HttpSession session = req.getSession();
+        String role = (String) session.getAttribute("role");
 
-        resp.sendRedirect("/?deleted=" + isDeleted);
+        if ("admin".equals(role)) {
+            Long id;
+            try {
+                id = Long.parseLong(req.getParameter("id"));
+            } catch (NumberFormatException e) {
+                throw new UserNotFoundException();
+            }
+            userDao.deleteUser(id);
+            logger.debug("Admin with ID="+ session.getAttribute("userId")+ " deletes user");
+
+            resp.sendRedirect("/adminPage");
+        } else {
+            req.setAttribute("message", "Ошибка. Войдите в систему снова.");
+            logger.debug("Access error");
+            RequestDispatcher dispatcher = getServletContext().getRequestDispatcher("/index.jsp");
+            dispatcher.forward(req, resp);
+        }
     }
 
     @Override
@@ -76,15 +104,25 @@ public class UserServlet extends HttpServlet {
         String name = req.getParameter("name");
         String password = req.getParameter("password");
 
-        Long id;
-        try {
-            id = Long.parseLong(req.getParameter("id"));
-        } catch (NumberFormatException e) {
-            throw new UserNotFoundException();
+        HttpSession session = req.getSession();
+        String role = (String) session.getAttribute("role");
+
+        if ("admin".equals(role)) {
+            Long id;
+            try {
+                id = Long.parseLong(req.getParameter("id"));
+            } catch (NumberFormatException e) {
+                throw new UserNotFoundException();
+            }
+
+            userDao.updateUser(new User(id, name, password));
+
+            resp.sendRedirect("/adminPage");
+        } else {
+            req.setAttribute("message", "Ошибка. Войдите в систему снова.");
+            logger.debug("Access error");
+            RequestDispatcher dispatcher = getServletContext().getRequestDispatcher("/index.jsp");
+            dispatcher.forward(req, resp);
         }
-
-        boolean isUpdated = userDao.updateUser(new User(id, name, password));
-
-        resp.sendRedirect("/?updated=" + isUpdated);
     }
 }
